enhance env, api

+++ app/src/main/java/com/ql2/myshop/domain/model/orders/OrdersModel.kt
@@ -5,6 +5,6 @@ import java.io.Serializable
 data class OrdersModel (
     val orderId: String? = null,
     val createdTime: String? = null,
-    val finalPrice: Float? = null,
+    val finalPrice: Int? = null,
     val orderStatus: String? = null
 ) : Serializable
